+++ andropilot/src/main/scala/com/geeksville/andropilot/ParameterListFragment.scala
@@ -10,43 +10,58 @@ import com.geeksville.flight.VehicleMonitor
 import android.app.FragmentManager
 import scala.collection.JavaConverters._
 import android.widget.SimpleAdapter
+import android.os.Handler
+import com.geeksville.flight.MsgParametersDownloaded
+import com.geeksville.util.ThreadTools._
 
-class ParameterListFragment extends ListFragment with AndroidLogger {
+class ParameterListFragment extends ListFragment with AndroServiceFragment {
 
-  private var vehicle: Option[VehicleMonitor] = None
+  override def onServiceConnected(s: AndropilotService) {
+    super.onServiceConnected(s)
 
-  override def onActivityCreated(saved: Bundle) {
-    super.onActivityCreated(saved);
+    // Don't expand the view until we have _something_ to display
+    debug("parameter list service connected")
+    updateParameters()
+  }
+
+  override def onVehicleReceive = {
+    // Super crufty - do this someplace else
+    case MsgParametersDownloaded =>
+      handler.post(updateParameters _)
+  }
+
+  private def updateParameters() {
+    // Don't expand the view until we have _something_ to display
+    if (getActivity != null) {
+      debug("updating parameters")
+      makeAdapter.foreach(setListAdapter)
+    }
   }
 
   override def onListItemClick(l: ListView, v: View, position: Int, id: Long) {
     info("Item clicked: " + id)
 
-    vehicle.foreach { v =>
-      val param = v.parameters(position)
-      val transaction = getFragmentManager.beginTransaction()
+    myVehicle.foreach { v =>
+      if (position < v.parameters.size) {
+        val param = v.parameters(position)
+        val transaction = getFragmentManager.beginTransaction()
 
-      var frag = new ParameterEditorFragment(param)
-      frag.show(transaction, "paramedit_fragment")
+        var frag = new ParameterEditorFragment(param)
+        frag.show(transaction, "paramedit_fragment")
+      }
     }
   }
 
-  private def makeAdapter() = {
-    // new ArrayAdapter(getActivity, android.R.layout.simple_list_item_1, vehicle.get.parameters)
-    val asMap = vehicle.get.parameters.toSeq.map { p =>
-      Map("n" -> p.getId.getOrElse("?"), "v" -> p.getValue.getOrElse("?").toString).asJava
-    }.asJava
-    val fromKeys = Array("n", "v")
-    val toFields = Array(R.id.param_name, R.id.param_value)
-    //new SimpleAdapter(getActivity, asMap, R.layout.parameter_row, fromKeys, toFields)
-    new SimpleAdapter(getActivity, asMap, R.layout.parameter_row, fromKeys, toFields)
-  }
+  private def makeAdapter() =
+    for (v <- myVehicle if !v.parameters.isEmpty) yield {
+      debug("Setting parameter list to " + v.parameters.size + " items")
 
-  def setVehicle(v: VehicleMonitor) {
-    vehicle = Some(v)
-    debug("Setting parameter list to " + v.parameters.size + " items")
-    // Don't expand the view until we have _something_ to display
-    if (v.parameters.size != 0 && getActivity != null)
-      setListAdapter(makeAdapter)
-  }
+      val asMap = v.parameters.toSeq.map { p =>
+        Map("n" -> p.getId.getOrElse("?"), "v" -> p.getValue.getOrElse("?").toString).asJava
+      }.asJava
+      val fromKeys = Array("n", "v")
+      val toFields = Array(R.id.param_name, R.id.param_value)
+      //new SimpleAdapter(getActivity, asMap, R.layout.parameter_row, fromKeys, toFields)
+      new SimpleAdapter(getActivity, asMap, R.layout.parameter_row, fromKeys, toFields)
+    }
 }